Implement JSON-LD to RDF conversion
This is needed as part of the input normalization for the DIVOC JWS
signatures. This is *far* from a complete or correct or spec-compliant
JSON-LD expansion and RDF conversion implementation, it just does barely
enough to produce the RDF output needed for the very specific purpose here.

While the RDF output is already correct for the proof sections, for the
content sections this still misses the URDNA2015 normalization algorithm
to resolve blank nodes correctly.