WHO ICD-11 code tree download, parsing and filtering
This will be used to generate data tables needed for ICAO VDS certificates.

This is rather messy, as the two download formats offered for this an XLSX
file and a Docker image with a query API (sic). We use the XLSX file,
LibreOffice to convert it to CSV and then rebuild the tree contained in
there. In that we can then select the sub-trees containing the disease and
vaccine codes.